LRQA-19336 Rename WebDriverHelper to WebDriverUtil so it's consistent with our usages of *Util
