+++ modules/test/poshi-runner/poshi-runner/src/main/java/com/liferay/poshi/runner/selenium/RetryWebElementImpl.java
@@ -15,6 +15,7 @@
 package com.liferay.poshi.runner.selenium;
 
 import com.liferay.poshi.runner.util.PropsValues;
+import com.liferay.poshi.runner.util.WebDriverUtil;
 
 import java.util.List;
 
@@ -467,7 +468,7 @@ public class RetryWebElementImpl
 		WebDriver webDriver = WebDriverSingleton.getWebDriver();
 
 		WebElement webElement = webDriver.findElement(
-			WebDriverHelper.getBy(_locator));
+			WebDriverUtil.getBy(_locator));
 
 		if (webElement == _webElement) {
 			System.out.println("Unable to find a new web element");
